Fix Login method
